% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcolumns.R
\name{addInsectWeatherCol}
\alias{addInsectWeatherCol}
\title{Add column of Insect x Weather interactions}
\usage{
addInsectWeatherCol(y)
}
\arguments{
\item{y}{}
}
\description{
Add column of Insect x Weather interactions
}

