% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcolumns.R
\name{addTempxPrecipColumns}
\alias{addTempxPrecipColumns}
\title{Add column of Temperature x Precipitation interactions}
\usage{
addTempxPrecipColumns(y)
}
\arguments{
\item{y}{}
}
\description{
Add column of Temperature x Precipitation interactions
}

