% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_function.R
\name{AIC_function}
\alias{AIC_function}
\title{Return table comparing AIC values for multiple models produced using SAS GLIMMIX.}
\usage{
AIC_function(y)
}
\description{
Return table comparing AIC values for multiple models produced using SAS GLIMMIX.
}

