% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_function.R
\name{BIC_function}
\alias{BIC_function}
\title{Return table comparing BIC values for multiple models produced using SAS GLIMMIX.}
\usage{
BIC_function(y)
}
\arguments{
\item{y}{}
}
\description{
Return table comparing BIC values for multiple models produced using SAS GLIMMIX.
}

