% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FruitPres_cAIC_function.R, R/cAIC_function.R
\name{cAIC_function}
\alias{cAIC_function}
\title{Return table comparing conditional AIC (cAIC) values for multiple models produced using SAS GLIMMIX.}
\usage{
cAIC_function(y)

cAIC_function(y)
}
\description{
Return table comparing conditional AIC (cAIC) values for multiple models produced using SAS GLIMMIX.

Return table comparing conditional AIC (cAIC) values for multiple models produced using SAS GLIMMIX.
}

