% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_function.R
\name{cleanAICModelSelectionZIP}
\alias{cleanAICModelSelectionZIP}
\title{Clean up AIC Model Selection for ZIP Models}
\usage{
cleanAICModelSelectionZIP(y, round.n = 2)
}
\arguments{
\item{round.n}{Number of digits to round to. Defaults to 2.}
}
\description{
First order by delta AIC, the model number, and vital rate. Then set the AIC model selection values to "" for the count vital rates.
}

