% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructConfInt.R
\name{constructConfInt}
\alias{constructConfInt}
\title{Construct Conf. Intervals}
\usage{
constructConfInt(parameter.estimates, round_n = 3)
}
\arguments{
\item{parameter.estimates}{Table of parameter.estimate output from SAS}
}
\description{
Construct Conf. Intervals
}

