% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructConfInt.R
\name{constructConfInt2}
\alias{constructConfInt2}
\title{Construct Conf. Intervals the Hard Way}
\usage{
constructConfInt2(parm.est)
}
\arguments{
\item{parm.est}{parameter.estimate output from SAS}
}
\description{
Construct Conf. Intervals the Hard Way
}

