% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructConfInt.R
\name{constructConfIntforWeatherData}
\alias{constructConfIntforWeatherData}
\title{Construct Conf. Intervals for Fecundity Random Effects}
\usage{
constructConfIntforWeatherData(y)
}
\arguments{
\item{parameter.estimates}{Table of parameter.estimate output from SAS}
}
\description{
Construct Conf. Intervals for Fecundity Random Effects
}

