% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertShorttoLong.R
\name{convertShorttoLong}
\alias{convertShorttoLong}
\title{Change from long to short format, using "Value"}
\usage{
convertShorttoLong(x)
}
\arguments{
\item{x}{}
}
\description{
Change from long to short format, using "Value"
}

