% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertShorttoLong.R
\name{convertShorttoLongX}
\alias{convertShorttoLongX}
\title{Change fit statistics from long to short format, changing cell value to either "X" or "."}
\usage{
convertShorttoLongX(x)
}
\arguments{
\item{x}{}
}
\description{
Change fit statistics from long to short format, changing cell value to either "X" or "."
}

