% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModelSelectionTable.R
\name{createModelSelectionTable}
\alias{createModelSelectionTable}
\title{Build table of candidate models}
\usage{
createModelSelectionTable(covariance.parameter.estimates, models.dimensions,
  convergence.status, parameter.estimates, conditional.fit.statistics)
}
\arguments{
\item{conditional.fit.statistics}{}
}
\description{
Build table of candidate models
}

