% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genmod_modelselection_format_parameter_estimates_function.R
\name{genmod_modelselection_format_parameter_estimates_function}
\alias{genmod_modelselection_format_parameter_estimates_function}
\title{Format parameters and their estimates for a model selection table using data from SAS GLIMMIX}
\usage{
genmod_modelselection_format_parameter_estimates_function(y)
}
\arguments{
\item{y}{}
}
\description{
Format parameters and their estimates for a model selection table using data from SAS GLIMMIX
}

