% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_results_processing_function.R
\name{model_results_processing_function}
\alias{model_results_processing_function}
\title{Process model results for a model selection table using data from SAS GLIMMIX models}
\usage{
model_results_processing_function(y, select_list)
}
\description{
Process model results for a model selection table using data from SAS GLIMMIX models
}

