% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_selection_table_function.R
\name{model_selection_table_function}
\alias{model_selection_table_function}
\title{Build table of candidate models}
\usage{
model_selection_table_function(covariance.parameter.estimates,
  models.dimensions, convergence.status, parameter.estimates,
  conditional.fit.statistics)
}
\arguments{
\item{covariance.parameter.estimates}{}

\item{models.dimensions}{}

\item{convergence.status}{}

\item{parameter.estimates}{}

\item{conditional.fit.statistics}{importFrom reshape2 dcast
importFrom dplyr mutate}
}
\description{
Build table of candidate models
}

