% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelselection_model_results_function.R
\name{modelselection_model_results_function}
\alias{modelselection_model_results_function}
\title{Format parameters and their estimates for a model selection table using data from SAS GLIMMIX}
\usage{
modelselection_model_results_function(models.dimensions, convergence.status,
  parameter.estimates, conditional.fit.statistics, select_list)
}
\arguments{
\item{models.dimensions}{}

\item{convergence.status}{}

\item{parameter.estimates}{}

\item{conditional.fit.statistics}{}

\item{select_list}{}
}
\description{
Format parameters and their estimates for a model selection table using data from SAS GLIMMIX
}

