% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_processing_function.R
\name{names_processing_function}
\alias{names_processing_function}
\title{Process model results for a model selection table using data from SAS GLIMMIX models}
\usage{
names_processing_function(y)
}
\description{
Process model results for a model selection table using data from SAS GLIMMIX models
}

