% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixed_model_selection_table_function.R
\name{nlmixed_table_function}
\alias{nlmixed_table_function}
\title{Build table of candidate models}
\usage{
nlmixed_table_function(models.dimensions, convergence.status,
  parameter.estimates, round.n = 2)
}
\arguments{
\item{conditional.fit.statistics}{}
}

