% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processClassLevels.R
\name{processClassLevels}
\alias{processClassLevels}
\title{Process ClassLevels output from SAS}
\usage{
processClassLevels(x)
}
\arguments{
\item{x}{}
}
\description{
Process ClassLevels output from SAS
}

