% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processColumnNames.R
\name{processColumnNames}
\alias{processColumnNames}
\title{Change Names of Columns in a Model Selection Table from SAS GLIMMIX Model Output}
\usage{
processColumnNames(y)
}
\arguments{
\item{select_list}{}
}
\description{
Change Names of Columns in a Model Selection Table from SAS GLIMMIX Model Output
}

