% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processColumnNames.R
\name{processModSelectionColumnNames}
\alias{processModSelectionColumnNames}
\title{Change Names of Model Selection Columns}
\usage{
processModSelectionColumnNames(y)
}
\arguments{
\item{y}{}
}
\description{
Change Names of Model Selection Columns
}

