% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processColumnNames.R
\name{processStatisticsColumnNames}
\alias{processStatisticsColumnNames}
\title{Change Names of Statistics Columns}
\usage{
processStatisticsColumnNames(y)
}
\arguments{
\item{y}{}
}
\description{
Change Names of Statistics Columns
}

