% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processColumnNames.R
\name{processT1P1ColumnNames}
\alias{processT1P1ColumnNames}
\title{Change Names of Precipitation and Temperature Columns}
\usage{
processT1P1ColumnNames(y)
}
\arguments{
\item{select_list}{}
}
\description{
Change Names of Precipitation and Temperature Columns
}

