% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModelSelectionTable.R
\name{processWeatherVariables}
\alias{processWeatherVariables}
\title{Paste PCA axis data together}
\usage{
processWeatherVariables(y)
}
\arguments{
\item{y}{}
}
\description{
Paste PCA axis data together
}

