% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randeffects_converg_status_function.R
\name{randeffects_converg_status_function}
\alias{randeffects_converg_status_function}
\title{Create table with the convergence status of multiple SAS GLIMMIX models}
\usage{
randeffects_converg_status_function(modelresults)
}
\arguments{
\item{modelresults}{}
}
\description{
Create table with the convergence status of multiple SAS GLIMMIX models
}

