% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randeffects_model_results_function.R
\name{randeffects_model_results_function}
\alias{randeffects_model_results_function}
\title{Merge summary statistics from SAS GLIMMIX models to compare random effects models.}
\usage{
randeffects_model_results_function(covariance.parms.test, convergence.status,
  covariance.parms.estimates, conditional.fit.statistics, fit.statistics,
  parms.estimates)
}
\arguments{
\item{covariance.parms.test}{}

\item{convergence.status}{}

\item{covariance.parms.estimates}{}

\item{conditional.fit.statistics}{}

\item{fit.statistics}{}
}
\description{
Merge summary statistics from SAS GLIMMIX models to compare random effects models.
}

