% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixed_model_selection_table_function.R
\name{replaceNLMIXEDnames}
\alias{replaceNLMIXEDnames}
\title{Replace Parameter Names for NLMIXED Output}
\usage{
replaceNLMIXEDnames(y)
}
\arguments{
\item{y}{Parameter estimates output}
}

