% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixed_model_selection_table_function.R
\name{replaceNLMIXEDnamesGTMNERR}
\alias{replaceNLMIXEDnamesGTMNERR}
\title{Replace Parameter Names for NLMIXED Output for GTMNERR O. stricta}
\usage{
replaceNLMIXEDnamesGTMNERR(y)
}
\arguments{
\item{y}{Parameter estimates output}
}

