% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixed_model_selection_table_function.R
\name{replaceNLMIXEDnamesOstricta}
\alias{replaceNLMIXEDnamesOstricta}
\title{Replace Parameter Names for NLMIXED Output for O. stricta}
\usage{
replaceNLMIXEDnamesOstricta(y)
}
\arguments{
\item{y}{Parameter estimates output}
}

