% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructConfInt.R
\name{roundpValues}
\alias{roundpValues}
\title{Round the p-values before constructing confidence intervals}
\usage{
roundpValues(y, pvalue_name = "Probt")
}
\arguments{
\item{y}{data frame}

\item{pvalue_name}{column name containing p-values. Default value is "Probt".}
}
\description{
Round the p-values before constructing confidence intervals
}

