% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_to_long_format_X_function.R
\name{short_to_long_format_X_function}
\alias{short_to_long_format_X_function}
\title{Change fit statistics from long to short format, changing cell value to either "X" or "."}
\usage{
short_to_long_format_X_function(x)
}
\description{
Change fit statistics from long to short format, changing cell value to either "X" or "."
}

