% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model_results_function_stricta.R
\name{survival_model_results_function_stricta}
\alias{survival_model_results_function_stricta}
\title{Create table of survival model results}
\usage{
survival_model_results_function_stricta(convergence.status, parameter.estimates,
  fit.statistics)
}
\description{
Create table of survival model results
}

