% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/X_function.R
\name{X_function}
\alias{X_function}
\title{Return table comparing SAS GLIMMIX random effects models.}
\usage{
X_function(x)
}
\arguments{
\item{modelresults}{}
}
\description{
Return table comparing SAS GLIMMIX random effects models.
}

