% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmod_modelselection_format_parameter_estimates_function.R
\name{genmod_modelselection_format_parameter_estimates_function}
\alias{genmod_modelselection_format_parameter_estimates_function}
\title{Format parameters and their estimates for a model selection table using data from SAS GLIMMIX}
\usage{
genmod_modelselection_format_parameter_estimates_function(y)
}
\description{
Format parameters and their estimates for a model selection table using data from SAS GLIMMIX
}

