% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dimensions_name_processing_function.R
\name{model_dimensions_name_processing_function}
\alias{model_dimensions_name_processing_function}
\title{Rename model dimensions}
\usage{
model_dimensions_name_processing_function(y)
}
\description{
Rename model dimensions
}

