% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelselection_model_results_function.R
\name{model_selection_results_function}
\alias{model_selection_results_function}
\title{Format parameters and their estimates for a model selection table using data from SAS GLIMMIX}
\usage{
model_selection_results_function(model_selection_table, select_list,
  round.n = 2)
}
\arguments{
\item{select_list}{}

\item{models.dimensions}{}

\item{convergence.status}{}

\item{parameter.estimates}{}

\item{conditional.fit.statistics}{}
}
\description{
Format parameters and their estimates for a model selection table using data from SAS GLIMMIX
}

