% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_table_function.R
\name{model_selection_table_function}
\alias{model_selection_table_function}
\title{Build table of candidate models}
\usage{
model_selection_table_function(covariance.parameter.estimates,
  models.dimensions, convergence.status, parameter.estimates,
  conditional.fit.statistics)
}
\arguments{
\item{conditional.fit.statistics}{}
}
\description{
Build table of candidate models
}

