% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/names_processing_function.R
\name{names_processing_function}
\alias{names_processing_function}
\title{Process model results for a model selection table using data from SAS GLIMMIX models}
\usage{
names_processing_function(y)
}
\arguments{
\item{y}{}

\item{select_list}{}
}
\description{
Process model results for a model selection table using data from SAS GLIMMIX models
}

