% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Process_PCA_Eigenvalues.R
\name{Process_PCA_Eigenvalues}
\alias{Process_PCA_Eigenvalues}
\title{Process PCA Eigenvalue Results from SAS}
\usage{
Process_PCA_Eigenvalues(PCA.all.surveys.eigenvalues, PCA.winter.eigenvalues,
  PCA.spring.eigenvalues)
}
\arguments{
\item{PCA.all.surveys.eigenvalues}{}

\item{PCA.winter.eigenvalues}{}

\item{PCA.spring.eigenvalues}{}
}
\description{
Process PCA Eigenvalue Results from SAS
}

