% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProcessClassLevels.R
\name{ProcessClassLevels}
\alias{ProcessClassLevels}
\title{Process ClassLevels output from SAS}
\usage{
ProcessClassLevels(x)
}
\arguments{
\item{x}{}
}
\description{
Process ClassLevels output from SAS
}

