% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection_converg_status_function.R, R/randeffects_converg_status_function.R
\name{randeffects_converg_status_function}
\alias{randeffects_converg_status_function}
\title{Create table with the convergence status of multiple SAS GLIMMIX models}
\usage{
randeffects_converg_status_function(modelresults)

randeffects_converg_status_function(modelresults)
}
\description{
Create table with the convergence status of multiple SAS GLIMMIX models

Create table with the convergence status of multiple SAS GLIMMIX models
}

