% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/short_to_long_format_function.R
\name{short_to_long_format_function}
\alias{short_to_long_format_function}
\title{Change from long to short format, using "Value"}
\usage{
short_to_long_format_function(x)
}
\arguments{
\item{x}{}
}
\description{
Change from long to short format, using "Value"
}

