% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model_results_function_humifusa.R
\name{survival_model_results_function_humifusa}
\alias{survival_model_results_function_humifusa}
\title{Create table of survival model results}
\usage{
survival_model_results_function_humifusa(convergence.status,
  parameter.estimates, fit.statistics, select_list)
}
\arguments{
\item{select_list}{}
}
\description{
Create table of survival model results
}

